requirements.txt
boto3>=1.26
python-dotenv>=1.0
pymongo>=4.0
waste-predictor>=1.0.0

